requirements.txt
adal==1.2.7
alembic==1.13.1
ansible-core==2.14.12
applicationinsights==0.11.10
argcomplete==3.1.6
Authlib==1.3.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-servicefabric==8.2.0.0
azure-storage-blob==12.19.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
docker==7.0.0
docker-image-builder==0.1.3
docker-run-build==0.1.0
docker-run-cmd==1.0.13
dockerfile-requirements==0.0.4
distlib==0.3.6
enum34==1.1.10
fastapi==0.110.0
filelock==3.13.1
Flask==3.0.2
Flask-Azure-Storage==0.2.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Register==0.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
forex-python==1.8
future==0.18.3
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
graphql-core==3.2.3
graphql-server-core==2.0.0
graphql-env==0.1.0
graphql_client==0.3.0
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.2
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
knack==0.6.3
lxml==4.9.1
Mako==1.3.2
MarkupSafe==2.1.5
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
msrestazure==0.6.4
mssql-scripter==1.0.0a23
oauthlib==3.2.2
packaging==23.2
paypalrestsdk==1.13.3
pillow==10.2.0
platformdirs==2.0.0
portalocker==2.8.2
psutil==5.9.6
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.4.2
PyJWT==2.8.0
pyodbc==5.1.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==5.4.1
requests==2.27.0
requests-oauthlib==1.3.1
resolvelib==1.0.1
rsa==4.9
sfctl==11.2.0
six==1.16.0
SQLAlchemy==2.0.27
starlette==0.37.1
stripe==8.5.0
tabulate==0.9.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.18
uvicorn==0.27.1
virtualenv==20.14.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
WTForms==3.1.2
yamlloader==1.1.0
zipp==3.17.0